Script file case2.mjs created
